Refactoring To Reduce Non Essential Duplication
